refactor: time based scheduler file structure
